Set Capacity =1000

+++ Kantar.AdExpress.Service/src/Kantar.AdExpress.Service.Core/Domain/UniversBranch.cs
@@ -25,7 +25,7 @@ namespace Kantar.AdExpress.Service.Core.Domain
         public long Id { get; set; }
         public int LabelId { get; set; }
         public string Label { get; set; }
-        public int Capacity { get; set; }
+        public long Capacity { get; set; }
         public string OverLimitMessage { get; set; }//2286
         public string SecurityMessage { get; set; }//2285
         public string ExceptionMessage { get; set; }//922
