+++ Kantar.AdExpress.Service/src/Km.AdExpressClientWeb/Models/MediaSchedule/MarketViewModel.cs
@@ -49,7 +49,7 @@ namespace Km.AdExpressClientWeb.Models.MediaSchedule
         public long Id { get; set; }
         public int LabelId { get; set; }
         public string Label { get; set; }
-        public int Capacity { get; set; }
+        public long Capacity { get; set; }
         public string OverLimitMessage { get; set; }//2286
         public string SecurityMessage { get; set; }//2285
         public string ExceptionMessage { get; set; }//922
